Split Memory Cache Implementation into Separate File.
